refactor : test 클래스 별로 분류
